move color formatter to own file
