+++ codex/logger/loggerd.py
@@ -1,41 +1,9 @@
 """Central Logging Thread."""
 import logging
 from logging.handlers import QueueListener, RotatingFileHandler
-from types import MappingProxyType
 
-from colors import color
-
+from codex.logger.formatter import ColorFormatter
 from codex.settings.settings import DEBUG, LOG_DIR, LOG_TO_CONSOLE, LOG_TO_FILE
-
-
-class ColorFormatter(logging.Formatter):
-    """Logging Formatter to add colors and count warning / errors."""
-
-    FORMAT_COLORS = MappingProxyType(
-        {
-            "CRITICAL": {"fg": "red", "style": "bold"},
-            "ERROR": {"fg": "red"},
-            "WARNING": {"fg": "yellow"},
-            "INFO": {"fg": "green"},
-            "DEBUG": {"fg": "black", "style": "bold"},
-            "NOTSET": {"fg": "blue"},
-        }
-    )
-
-    def __init__(self, fmt, **kwargs):
-        """Set up the FORMATS dict."""
-        super().__init__(**kwargs)
-        self.formatters = {}
-        for level_name, args in self.FORMAT_COLORS.items():
-            levelno = getattr(logging, level_name)
-            template = color(fmt, **args)
-            formatter = logging.Formatter(fmt=template, **kwargs)
-            self.formatters[levelno] = formatter
-
-    def format(self, record):
-        """Format each log message."""
-        formatter = self.formatters[record.levelno]
-        return formatter.format(record)
 
 
 class CodexLogQueueListener(QueueListener):
